+++ spec/components/error_message_spec.rb
@@ -11,6 +11,12 @@ describe "Error message", type: :view do
     assert_select(".govuk-error-message", text: "Error: Please enter your National Insurance number")
   end
 
+  it "renders error message with a given ID" do
+    render_component(text: "Please enter your National Insurance number", id: "error_id")
+
+    assert_select(".govuk-error-message[id=error_id]")
+  end
+
   it "escapes HTML in items and allows HTML safe strings" do
     render_component(items: [
       { text: "Error where HTML is <strong>escaped</strong>" },